move RoundPhaseState struct to models.
