+++ WalletWasabi.Gui/Converters/PhaseStringConverter.cs
@@ -1,11 +1,9 @@
 using Avalonia.Data.Converters;
 using System;
-using System.Collections.Generic;
 using System.Globalization;
-using System.Text;
 using WalletWasabi.CoinJoin.Common.Models;
 using WalletWasabi.Exceptions;
-using WalletWasabi.Gui.Controls.WalletExplorer;
+using WalletWasabi.Gui.Models;
 
 namespace WalletWasabi.Gui.Converters
 {
